create location component and move things into crossing list item folder
