+++ frontend/src/Dashboard/CrossingListPage/CrossingListItem/Location.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import FontAwesome from 'react-fontawesome';
+import './Location.css';
+
+class Location extends React.Component {
+  render () {
+    return (
+      <div className="Location">LOCATION</div>
+    );
+  }
+}
+
+export default Location;
