+++ frontend/src/Dashboard/CrossingListPage/CrossingListItem/Location.scss
@@ -1,0 +1,5 @@
+@import "../../../css/flexbox";
+
+.Location {
+  @include flexitem;
+}
