+++ frontend/src/Dashboard/CrossingListPage/CrossingListItem/StatusToggle.scss
@@ -1,5 +1,5 @@
-@import "../../css/flexbox";
-@import "../../css/font-awesome/font-awesome.scss";
+@import "../../../css/flexbox";
+@import "../../../css/font-awesome/font-awesome.scss";
 
 $container-height: 46px;
